requirements.txt
python-telegram-bot==12.7
matplotlib==3.1.2
antlr4-python3-runtime==4.8
